Refactor LLMServer initialization and thread handling; update service name to HTTPServer and use constants for port and thread counts

+++ include/llmserver.hpp
@@ -44,7 +44,7 @@ private:
     HANDLE vStopEvent;
     HANDLE vIOCompletionPort;
 
-    DWORD vThreadCount;
+    WORD vThreadCount;
     std::vector<std::thread> vThreads;
     ULONG vConnectionContextCount;
